agents/retriever_agent/requirements.txt
langchain
openai
faiss-cpu
tiktoken
uvicorn
fastapi
sentence-transformers
